temporary about
this is just the code for the about page with the working back to top
button, I have to commit this new page in order to be able to save my
code, because github does not let me synch otherwise (saying my existing
code will be overwritten). this should not imply any change to any other
page.