fixed links to burger menu files
I just fixed the links in the index and history pages as the burger menu
css and js files have to be in the css and js folders respectively  (I
don't know how they got out of there!). Menu also became a bit more
readable (less transparenty background) on charged backgrounds such as
the pictures in moods and history. I am open to a change in background
color though  (it could be grey) and not too convinced by the helvetica
font.